WPs
Fixed work_packages function to include the user in the daily_record not
db.getuid () (the current user). This gives the list of wps for the user
of the daily record not the one (e.g. controlling) editing a record.
Added all wps with empty allow list to the list of allowed wps.