Bug-fix
Onchange of date field we want to reload, even if editing is not
allowed. We also reset the default action to the default in the hidden
value.
Save button now does not use default submit action, because a previous
try to submit to supervisour could have changed the action, then all
editing of fields (and the save button) would try to submit to
supervisor instead of just saving.