SMB
Update smb_domain with last_gid, last_uid, last_machine_uid infos.
org_location added to smb_domain.
machine_group in smb_domain is now a Number instead of a Link to group,
the group needs an org_location and we want to check the gid in new
groups for consistency with the org_location->>smb_domain -- so we have
to first define the machine gid in smb_domain then add the group.
page.html default_query update.