New overtime
Duration now gets __str__ and __repr__ methods for debugging, for this
it gets the user and date.
Remove dead code in Period_Data (which already had an assert(0)).
Bug-fix in compute_running_balance: This had an off-by-one error with
sharp_end, resulting in a balance of the last day of an interval to move
to the next period if it happened to be the first day of a new period.
More commented-out debug code in compute_running_balance.
Remove last FIXME in new regressin test.