Fix schema definition for time_ldap
Add min_adr and contact.
Fix security settings for user_contact to include IT role only if
defined.
Factor security settings for contact into contact_sec and import the
latter in all relevent schemadefs.
Add nosy role definition to doc.py as it is a nosy class and uses the
role.