Implement new features for IT-Tracker
Closed issues stay now closed if not changed by role IT.
New message on closed issue would re-open the issue but due to
closed-check this does happen only for role IT -- except if the status
of the issue is set in the same transaction anyway.
Setting duplicate (superseder) will now set required attributes
responsible (not helpdesk) and it_prio (not unknown) if necessary.
This ensures that the issue is really closed.