Add sync of new guid attribute
Now matching of ldap users is done preferrably via the guid not via the
username. This way we can now change usernames in ldap and (after an
initial sync of all guids) will correctly update the user in roundup
changing the username.
Also allow display of (hex-encoded) guid value in user searches.