Update ldap sync to not change roles
We need to first sync all (possibly renamed) users from ldap and *then*
in a second round all remaining roundup users. Otherwise it could happen
that a user is first marked obsolete (because renamed in ldap and not
found when syncing the roundup user) and *then* again found via ldap and
resurrected. In that case roles have been removed and the user might
lose roles in the process.