Fix several problems
In lib/common.py we had a forgotten debugger call and an assertion that
failed due to rounding, in lib/schemacfg/pr.py we had a check method
that looked for stati that might not be in the database. And finally in
detectors/pr.py a variable name shadowed a global function.