Fix permissions, allow re-open
Permissions to edit nosy and files updated: Anybody on nosy or approval
list may now edit nosy and files.
Now it is possible to re-open an issue from status rejected. We clear
all approvals in this case and start anew.
Creator is now put on the nosy list only on initial submit so he/she can
remove him/herself later.