Fix customer default fromaddress
The default from address (for support issue emails *to* this customer)
used to be the support-nosy address. It is now taken from the "To" field
of the email which should be one of the support addresses. To be sure we
check the mail address begins with 'support'. If no address is found (or
it doesn't start with 'support') we now have a configurable
'support_from' config-item in the mailgw section of the extended config.