Distinguish CC and To in customer mails
Add a new attribute cc_emails to select emails to customer that will be
placed on CC. Now Email headers are correctly generated: All CC emails
will be put on CC:, Mails from 'emails' field will be put on To. The To
field can be empty, one of CC and To must be specified.