Fix parsing of CC and To field
.. on incoming support issue.
Now we do *not* put the support@ address in CC if we detect that it is
the fromaddr of the customer.
We distinguish CC and To: Addresses from To are put in emails, addresses
from CC are put in cc_emails.
All the rest (mails that are unknown) are put on CC.
We now also try to match incoming mail addresses (from To and CC)
against the maildomain of the current customer (if not on the list of
the customer). If it matches, the email is appended to the customer
list.