Fix retirement when syncing
Now we detect local nodes that are not referenced by the remote. We now
try to match time_project via the description. Small fix for looking up
a list for a multilink: This used to fail if *one* item could not be
looked up (e.g. purchasing agent). This now fails only if *no* item can
be found. This reduces the list of items linked via a multilink to the
ones being still valid.