Make board and finance approvals configurable
Both are no longer hard-coded, the role name can change. We also allow
to configure approvals with a list of organisations, so we can have
different finance/board approvals (and other) by organisation.

This removes the max_sum property from currency: This is now configured
in the board approval.

Note that there are now exceptions if no board/finance approval is found
for the current PR.