Factor clearance_by
Now we have tt_clearance_by (which uses the supervisor to check if we
may approve time records) and approval_by which can be used for other
purposes with the same approval delegated/substitute/substitute active
mechanism. Also add a new possible attribute subst_until that replaces
subst_active with a date.