Fix purchasing agent computation
Now we remove agents that do not have view permission on *all* involved
purchase types: We now also check purchase types of offer items.
We force a check whenever relevant properties of an offer item change.
The fallback display of purchasing agents in the purchase request form
is gone: We no longer display the purchasing agent from the sap_cc or
time category in case the agent field is empty.
Also fix require_attributes computation for multilinks: They should not
be checked for None but for emptyness.