Check vacation changes
Allow certain vacation changes only with an absolute vacation
correction. In particular the vac_aliq may not change without a vacation
correction. Also the yearly vacation may not change without a correction
if Daily alliquotation is in use.
When we're at it fix that vac_aliq is copied when copying user_dynamic
in the web interface.