Fix vacation duration
Add new vacation routine to fix vacation durations for a given date
range, by default (no range given) we fix all non-frozen records.
This is needed if the work hours in the dynamic user record are changed
and/or a new record is added with different times.
In addition we provide a script to fix the durations for a given user
(for wrong records that exist before this change).