Vacation: fix special case of re-employment
Check for an absolute vacation correction when starting a new employment
of someone who was already employed earlier: An absolute vacation
correction must exist *before* (not *at*) the date of the start of the
new dynamic user record and this correction must be *after* or *at* the
end-date of the last dynamic user record.

This way the vacation correction is valid for the new employment and the
check which was too restrictive now doesn't fail.