Make a summary report redirect configurable
Now if in section [ttt] in extension/config.ini the variable
summary_report_redirect is set to a redirect link, the old summary
report is shown as "Summary report (old)" and a new redirect link is
added (which points to a template with meta magic for the redirect)
under the title Summary Report. If the variable is not set we get the
old behavior.