Allow creation of new dynuser record with valid_to
Until now only creation at the end was supported. Now we allow creation
of a new dyn user record at the end of the period of an existing record.
The new record must have exactly the end time of the record that is
split and it must have a start time later than the original start time.