Fix problem in flexi_remain
This would call flexitime_submission_days with a negative date range
(begin > end) if no dynamic user record exists for the start of the
year.

This in turn would call leave_submission_days which threw an assertion,
leave_submission_days now has an assertion that begin <= end which will
immediately show that the caller violates the precondition should this
happen again.