Fix domain permissions on user
Now the domain edit permission allows Edit *and* View permission,
previously only View was given (and resulted in some user properties not
visible in the API). Also these permissions are now tied to the domain in
the user record, so users without that domain will not be visible (except
for the things everybody may see) to the domain edit users.