Tolerate duplicate contacts in LDAP
For contacts (phone numbers etc.) a duplicate would continuously
re-create the duplicate item in roundup. The reason is that a dictionary
is used during the sync and when the duplicate is synced the item is
already removed from the dictionary. We now use a second dictionary with
sunced keys which has the beneficial side-effect of removing duplicates.