Fix ldap detectors for user_contact
Do not check domain_user_check on creation: user_contact records are
created without a user and later added to the list of contacts in user.
So we must allow creation unconditionally.
Currently on modification of user_contact no ldap update is triggered,
this results in the change being lost. Generalize sync_to_ldap_ud_c and
call it for user_contact, too.