Add reverse multilinks
Now we have a rev_multilink to wps in time_project and from user to
user_functional_role. It's now possible to search in the reverse
direction, e.g. all users with some functional role. The default view
for timecat is now all timecats where the user is on the bookers list of
one of the wps.
Also fix the case that user_functional_role wasn't editable: The default
templating generates a required property (@required) that contains the
labelprop. The default labelprop was activity (aka Last Activity). The
default labelprop is now the functional role and we require all
properties, namely, the functional role, the user and the ratio.