Fix ldap sync issues for issue-tracker
Allow that now user_status with name 'valid' exists, we search for
something *containing* the string 'valid' if a lookup fails.
For the ldap sync we now look up (in roundup) the user without the
domain name if we do not find a matchin user including the domain.
The user_create_magic isn't called if the class org_location doesn't
exist.