Fix sequencing bug in PR editing
When approving a PR with an additional message, the sequence of events
(when the message was appended to the PR vs. when the pr_approval was
processed) was arbitrary. This is because there were no dependencies
between the PR and the approvals. Now if the PR was changed *after* the
approvals it would remove messages and nosy-list changes done by the
approval reactor. This again because the old values of the PR are
computed *before* the whole sequence of events in the EditItemAction.
So changes done by reactors of other classes are not yet seen at this
point.  So we add a link update from the approval to the PR (the link is
already present and won't change but this *will* order the approvals
after the PR).

Fixes Request #17632961:
"Infosec approver werden automatisch von nosy geloescht"