Fix vacation correction check for user creation
During user creation a dynamic user record is created automatically which
does also trigger the creation of an initial absolute vacation correction.
This vacation correction should find the first dynamic user record that
is created within this process and allow creation if the vacation correction
if created for the same year as the dynamic user record.