Fix room sync
The check when syncing to roundup did not correctly check the
configuration: For some items we want to sync back to roundup only if
there is a valid vie_user link to this user (in that case attributes
previously synced to LDAP via the vie_user backlink will be synced to
roundup in the next round). There is already a config item for this case
but the check was missing to verify that there is a backlink.

This bug is probably also relevant for other attributes that should be
synced back to roundup *only* if there is a corresponding vie_user_ml
link.