Fix user_dynamic check when there is a vie_user
.. and fix tests so they do not fail.
Use unicode literals in test and convert to str where necessary,
this should pass for python3 in the future and gets rid of lookup
problems with ldap (which returns unicode).
Unfortunately the number of LDAP updates is now reduced by one because
one user doesn't get updates (because now a unicode compare yields a
match). This changes the number of LDAP changes in a message in three
tests.