Add new picture sync
Configurable size of synced picture and configurable quality.
Pictures are *always* converted to jpeg when writing to LDAP.
See extensions/config.ini.example in the limit section for the two new
options picture_sync_size and picture_quality.
We now use a smaller default size because the cloud services have lower
size limits.
Note that the regression test uses a picture from wikimedia which is
checked-in as test/240px-Bald_Man.svg.png -- the original is from
https://commons.wikimedia.org/wiki/File:Bald_Man.svg -- we're using the
240x240 pixel version which is about 15k.
Also note that the regression test of the resulting synced picture may
be dependent on the PIL version (python imaging lib) in use. It's
probably a good strategy to support several PIL versions in the
regression tests. Currently tested with Pillow-6.2.2.