Add automatism for public holidays
Now a public holiday time_record is added when a daily record is created
where a public holiday exists
Updates are performed when the public holiday changes (is_half flag is
changed) or when the dynamic user record is change. These changes work
even if the time record is in status submitted/accepted or leave. The
leave is also corrected if the is_half flag of the public holiday
changes.
Note that there is currently no special handling when a public holiday
is added or retired. This simply should not happen...