Fix bug that attendance record is left over
This happened when the duration is empty (deleting the time record) but
the attendance start/end was also empty. In that case nothing happens to
the attendance record. We now set start/end of the attendance record to
the same (non-empty) value so that the attendance record is also
deleted. We cannot solve this in a reactor due to different code paths:
In one path the ar may be deleted first while in another path the tr may
be deleted first depending on dictionary order of the request
dictionary.