Fix change of work package in accepted leaves
Change of WP in accepted leaves happens with the web interface if the
existing leave WP is no longer bookable (due to auto-wp changes). It
also happens to existing time records when public holidays are updated
and the WP is no longer valid. Both cases now have exceptions with
checks under which circumstances the change is allowed.