Bug-Fix in leave approval mail
When the user enters a leave submission that would consume more vacation
than they have, the actual requested vacation was not taken into account
when computing the remaining vacation. So in many cases we would send
the wrong email (to the supervisor, not to the supervisor *and*
hr_leave_approval role, and also with the wrong text). The computation
explicitly excempted the current vacation from the sum, maybe because in
the (distant) past submitted leaves were considered in the vacation sum.