Implement checks for linked items in PR
Now depending on the requester we may not link to items where the
allowed organisations of the requester are not matched in the linked
items.
Fix search permissions for pr_supplier_{rating,risk}.
Fix searching and display/creation of organisation in o_permission.
In help-popups for sap_cc and psp-element in PR: show organisation.
Fix empty set in organisation/org_location computation in o_permission
auditor.