Fix problem with permission error
This fixes the following sequence: When a user tries to cancel a PR and
forget the (mandatory) message the page is in a state where on the next
submit a permission error occurs.
This is originally documented in #50545664 and fixes the original
request (the issue now has been morphed into a feature request to permit
state change from rejected to cancelled).