Fix 'create similar' link for PR
Due to glitches in the Sync from the Timetracker, we occasionally have
retired SAP-CC and/or PSP elements, we now try to look up a newer
version of the same SAP-CC/PSP-element for the 'create similar' link.
Internal order is special because of the numeric name, this is now also
correctly filled in. G/L account was added, same for delivery address.
PR justification is left out intentionally. Fixes #54918780.