Fix problem with duplicate cursor
And go over the display of the first non-frozen daily record, we need a
redirect to make the next/prev buttons work correctly.
The real fix for the cursor problem would be to correctly handle named
cursors in the postgres backend (which is the only one using named
cursors currently).