Add test for special case in vacation computation
A dyn user record might have been modified to change the days a person
needs to work. This results in the duration of a time record to become
zero. This case is OK, we do not need to ensure that the work hours on
that day are non-zero (instead we ensure it *is* zero).