Fix formatting of offer items
New formatting of price and sum: both get a space for each thousand, the
price is moved into the last item of the offer item sum.
Add pr_offer_item_formatted for formatting a sum with spaces.
Add pr_offer_item_sum_formatted which is used in HTML and directly
produces the formatted output for the sum.
Move gl_account into the prior-to-last item in an offer item.
Remove commented-out old table layout.
Fix bug in formatting of preliminary approvals: This used
pr-section-approval instead of with -approvals (with s) resulting in
rose background in the test setup.
Stylesheet: pr-section is now a flex container and pr-section-fullwidth,
pr-section-offeritem use rounded borders.
New pr-item-sum for aligning the price in each item.