Add script to fix sap-cc in PRs
Due to sync problems some sap-ccs have been retired and re-created
during sync. So we find PRs where the sap-cc is retired *and* where we
can find a current sap-cc with the same name. If so we replace the
sap-cc (in the future, currently we only print out the cases).